Refactoring unit test cases so each *.cpp file is it's own compilation unit (now it's cleaner and easier to add new unit tests)
