+++ test/src/test_callback.cpp
@@ -3,6 +3,15 @@
  */
 
 /**************************************************************************************
+ * INCLUDE
+ **************************************************************************************/
+
+#include <catch.hpp>
+
+#include <TestUtil.h>
+#include <ArduinoCloudThing.h>
+
+/**************************************************************************************
  * GLOBAL CONSTANTS
  **************************************************************************************/
