+++ test/src/test_publishEvery.cpp
@@ -2,6 +2,19 @@
  * Copyright (c) 2019 Arduino.  All rights reserved.
  */
 
+/**************************************************************************************
+ * INCLUDE
+ **************************************************************************************/
+
+#include <catch.hpp>
+
+#include <TestUtil.h>
+#include <ArduinoCloudThing.h>
+
+/**************************************************************************************
+ * TEST CODE
+ **************************************************************************************/
+
 SCENARIO("A Arduino cloud property is published periodically", "[ArduinoCloudThing::publishEvery]")
 {
   /************************************************************************************/
