+++ test/src/test_publishOnChangeRateLimit.cpp
@@ -2,6 +2,19 @@
  * Copyright (c) 2019 Arduino.  All rights reserved.
  */
 
+/**************************************************************************************
+ * INCLUDE
+ **************************************************************************************/
+
+#include <catch.hpp>
+
+#include <TestUtil.h>
+#include <ArduinoCloudThing.h>
+
+/**************************************************************************************
+ * TEST CODE
+ **************************************************************************************/
+
 SCENARIO("A Arduino cloud property is published on value change but the update rate is limited", "[ArduinoCloudThing::publishOnChange]")
 {
   /************************************************************************************/
