+++ test/src/test_readOnly.cpp
@@ -2,6 +2,19 @@
  * Copyright (c) 2019 Arduino.  All rights reserved.
  */
 
+/**************************************************************************************
+ * INCLUDE
+ **************************************************************************************/
+
+#include <catch.hpp>
+
+#include <TestUtil.h>
+#include <ArduinoCloudThing.h>
+
+/**************************************************************************************
+ * TEST CODE
+ **************************************************************************************/
+
 SCENARIO("A Arduino cloud property is marked 'read only'", "[ArduinoCloudThing::decode]")
 {
   /************************************************************************************/
